+++ net-deleter/src/handlers/mod.rs
@@ -1,2 +1,3 @@
+pub mod clear_buffer_handler;
 pub mod delete_network_handler;
 pub mod delete_packet_handler;