+++ net-reporter/src/handlers/filters_handlers/network_overview/handler.rs
@@ -31,7 +31,7 @@ FROM
             SUM(Packet_Length) AS Total_Bytes,
             STRING_AGG(Protocols, ':' ORDER BY Protocols) AS Concatenated_Protocols
         FROM Network_Overview_Filters_Materialized_View
-        WHERE Tenant_ID = $1 AND Frametime >= $2 AND Frametime < $3
+        WHERE Tenant_ID = $1 AND Frametime >= $2 AND Frametime < $3 AND Network_ID = $4
         GROUP BY Src_IP
     ) AS lhs FULL OUTER JOIN (
         SELECT
@@ -39,7 +39,7 @@ FROM
             SUM(Packet_Length) AS Total_Bytes,
             STRING_AGG(Protocols, ':' ORDER BY Protocols) AS Concatenated_Protocols
         FROM Network_Overview_Filters_Materialized_View
-        WHERE Tenant_ID = $1 AND Frametime >= $2 AND Frametime < $3
+        WHERE Tenant_ID = $1 AND Frametime >= $2 AND Frametime < $3 AND Network_ID = $4
         GROUP BY Dst_IP
     ) AS rhs ON lhs.IP = rhs.IP;
 ";
@@ -57,11 +57,13 @@ impl NetworkOverviewFiltersHandler {
         tenant_id: &str,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
+        network_id: &str,
     ) -> Result<Vec<FilterEntryResponse>, Error> {
         sqlx::query_as(NETWORK_OVERVIEW_FILTERS_QUERY)
             .bind(tenant_id)
             .bind(start_date)
             .bind(end_date)
+            .bind(network_id)
             .fetch_all(connection_pool.as_ref())
             .await
     }
@@ -82,12 +84,13 @@ impl NetworkServiceHandler for NetworkOverviewFiltersHandler {
         let request = NetworkOverviewDashboardFiltersRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
-
+        let network_id = request.get_network_id();
         let executed_query_response = Self::execute_query(
             connection_pool,
             tenant_id,
             request_start_date,
-            request_end_date
+            request_end_date,
+            network_id,
         ).await?;
 
         let response: NetworkOverviewFiltersResponse = executed_query_response.into();